Fix mixin annotations, add nullability suppressions, add missing translation keys, fix deprecated API usage
- Add @Unique annotations to mixin methods in FrogportBlockEntityMixin, BasinOperatingBlockEntityMixin
- Fix parameter names in FrogportBlockEntityMixin (tag), FrogportRendererMixin (buffer)
- Add @SuppressWarnings("NullableProblems") to render layers and codecs
- Add missing translation key: key.categories.amphibia
- Fix deprecated builtInRegistryHolder() usage in EnderFrogTeleportGoal
- Register FrogportBlockEntityMixin in amphibia.mixins.json

+++ src/main/java/com/sanhiruzu/amphibia/block/GeneticFrogspawnBlock.java
@@ -46,7 +46,7 @@ public class GeneticFrogspawnBlock extends FrogspawnBlock implements EntityBlock
                 tadpole.setPersistenceRequired();
                 double d0 = (double)pos.getX() + random.nextDouble();
                 double d1 = (double)pos.getZ() + random.nextDouble();
-                tadpole.moveTo(d0, (double)pos.getY(), d1, random.nextFloat() * 360.0F, 0.0F);
+                tadpole.moveTo(d0, pos.getY(), d1, random.nextFloat() * 360.0F, 0.0F);
                 level.addFreshEntity(tadpole);
             }
         }
